refactor: merged (maps)/maps pages into main maps folder.
