+++ .changeset/lemon-laws-rule.md
@@ -1,0 +1,5 @@
+---
+"geohub": patch
+---
+
+refactor: merged (maps)/maps pages into main maps folder.
