--- sites/geohub/src/routes/(map)/maps/[id]/edit/+page.svelte
@@ -1,8 +1,0 @@
-<script lang="ts">
-	import Map from '$components/pages/map/Map.svelte';
-	import type { PageData } from './$types';
-
-	export let data: PageData;
-</script>
-
-<Map bind:defaultStyle={data.config.DefaultMapStyle} />
