--- sites/geohub/src/routes/(map)/maps/edit/+page.ts
@@ -1,11 +1,0 @@
-import type { PageLoad } from './$types';
-
-export const load: PageLoad = async () => {
-	const title = 'New Map | GeoHub';
-	const content = 'Create a new map';
-
-	return {
-		title,
-		content
-	};
-};
